fuzz: oneshot explicit close + notify poll interleave (#160, #161)
Two new structure-aware fuzzers for sync primitive race conditions:

1. notify_one_poll_interleave.rs (Tick #160): Tests notify_one()
   interleaved with notified.poll() patterns. Validates poll() only
   returns Ready after notify_one(), never spuriously. Includes
   spurious readiness detection and comprehensive operation coverage.

2. oneshot_explicit_close.rs (Tick #161): Tests explicit close via
   sender/receiver drops with subsequent send/recv operations.
   Validates post-close send returns Disconnected/Cancelled and
   post-close recv returns Closed/Cancelled with proper state consistency.

Both include comprehensive invariant tracking, concurrent testing modes,
and edge case coverage for race condition detection.

+++ fuzz/fuzz_targets/oneshot_explicit_close.rs
@@ -18,12 +18,9 @@ use asupersync::channel::oneshot;
 use asupersync::cx::Cx;
 use asupersync::util::ArenaIndex;
 use asupersync::types::{RegionId, TaskId, Budget};
-use std::sync::atomic::{AtomicUsize, AtomicBool, Ordering};
-use std::sync::Arc;
+use std::sync::atomic::{AtomicUsize, Ordering};
 use std::collections::HashMap;
-use std::task::{Context, Poll, Waker};
-use std::pin::Pin;
-use std::future::Future;
+use std::task::Waker;
 
 #[derive(Debug, Clone, Arbitrary)]
 struct OneshotCloseConfig {
@@ -186,8 +183,8 @@ impl TrackedChannel {
     fn new(channel_id: u8) -> Self {
         let (sender, receiver) = oneshot::channel::<i32>();
         let cx = Cx::new(
-            RegionId::from_arena(ArenaIndex::new(0, channel_id as u16)),
-            TaskId::from_arena(ArenaIndex::new(0, channel_id as u16)),
+            RegionId::from_arena(ArenaIndex::new(0, channel_id as u32)),
+            TaskId::from_arena(ArenaIndex::new(0, channel_id as u32)),
             Budget::INFINITE,
         );
